Refactor AuthService and UserService to utilize ErrorCodes for exception handling
- Updated AuthService to replace MessageConstants with ErrorCodes for various exceptions, enhancing clarity and maintainability.
- Refactored UserService to align with the new ErrorCodes structure for user-related exceptions.
- Introduced new error codes in ErrorCodes.java for email verification and password reset scenarios.
- Improved exception messages for better consistency across the authentication module.

+++ BE/common/src/main/java/app/together/common/workflow/entity/CoinPackage.java
@@ -35,7 +35,7 @@ public class CoinPackage extends BaseAuditEntity {
     Integer bonusCoins;
 
     @Column(name = "price_vnd", nullable = false, precision = 10, scale = 2)
-    BigDecimal priceVnd;
+    Long priceVnd;
 
     @Column(name = "is_popular")
     Boolean isPopular;
